Small improvements to find_arithmetic_expr
Return type for static type checking

Argument type reverted to UInt8 bcos the recursive solution doesn't seem likely anymore.

Added rationale for going brute force in comment.

Changed function docstring to comment because this belongs as an internal function, not as a documented external interface.

And the beginnings of an iterative collecting-and-operating way of finding the solution.